redis: fix memory leak
after a bunch of digging i realized that we were leaking instances
of redis pipelines that were never getting cleaned up. the cleaner
way of handling this is to only patch the classes and we can
avoid the issue entirely

i added a test that reveals the issue but automating it was very
difficult. the manual test will suffice for now.