add base context to the current tracing system
We introduce contexts. A contexts store a hierarchy of spans; this way,
it is possible to have multiple traces in parallel, as long as each
component keeps track of its context.

The tracer contains a default tracer which is used for all traces
created without the Context argument. This means that the existing code
continues to work the same way without any modification.