[tests] Update docker compose file version and images tags (#585)
* Update the docker compose file version

Moved to 3. I did not put the latest version so we are more lenient
about the various docker envs that users may have. As we need new
feature we can increse the file version accordingly.

* Update docker compose image versions

The following process has been adopted to decide whether to advance or not a version:
- if a new version exists that does not introduce breaking changes based on online info and change log, then advance.
- if an official '-alpine' version exists, then use it.

* Avoid using the .env file for docker compose env configuration

We had a few envs in the .env file that were used to configure postgres
and mysql credentials. As a matter of fact, they were basically useless.
Their values have been hard-coded in docker-compose.yml. I made sure that
they are the same values as in 'tests/contrib/config.py' as defaults.

Users can still override mysql and postgres credentials, e.g. to use a local
running version of the database, if the desire setting the envs as per
'tests/contrib/config.py'