writer: fix a possible race condition in put() if queue is full
In this new episode of "Multithreading is Hard", let's a fix a rare race
condition that could occur if a thread emptied the queue just before another
thread tries to replace a random item (because the queue is full).

In order to avoid that, we check the size of the queue, and simply retry to put
the item if the queue has been emptied in the meantime.

This also adds a tiny test for the `get` method.