tests/aiohttp: yield back to the loop until complete
The current code might fail because the loop is only executed once; if one of
the thread is late, its request won't be processed.

The call to `t.join()` has a timeout and has basically no effect here. It's now
moved at the end of the waiting loop to be sure all threads are completed.

Since pytest would not catch an assertion failure in a thread, the assertion
check for replies is moved to the main thread.

Finally, check the number of spans rather than the number of trace as they are
no reason we could have only one trace.

Fixes #876