span: use system random source to generate span id
The currently used random generator is not re-seeded on `os.fork` calls, making
span id collision a sure thing.

This is only a problem on Python 2 — Python 3 does not show this issue.

This patches changes the random number generator to use the system ones based
on urandom. This avoids the issue altogether.