Use setuptools_scm to handle version numbers
This leverages setuptools_scm to generate version number for the libary rather
than updating manually. It will leverages git tags to produce adequate numbers.

For dev version, it'll generate version numbers such as
`0.27.1.dev7+gdeb1fb54.d20190723` which are pretty explicity about where the
version is at.

Since we now produce and distribute wheels, setuptools-scm is only a
requirement if someone installs from source — which should not be the case for
anyone using pip.

This should simplify maintenance and distribution overall.