Create test for empty middleware (#1022)
* Create test for empty middleware

When a user declares an empty list/tuple for MIDDLEWARE on newer
versions of Django, the MIDDLEWARE setting should still be injected with
the ddtrace values.

Deliberately defined in a separate class to avoid the patch/setup that
happens in the setUp for the current class definition. For overrides of
the default settings, this must use the self.settings context manager
for restricting the change during the patching.

* Only fallback to MIDDLEWARE_CLASSES if MIDDLEWARE is None

This allows MIDDLEWARE to support falsy values like `[]` or `tuple()`
which are valid declarations for a Django application that has no
middleware.