fix(profiling): fix negative CPU time (#1437)
* feat(periodic): add support for thread native id in gevent threads

Python ≥ 3.8 supports native OS thread id. Assign it to gevent custom thread
too.

* fix(profiling/stack): use native thread id as CPU time identifier

The current code uses the pthread_t id to cache CPU time between two polling.
This is very optimistic as the pthread_t id can be reused, and it seems Linux
does that quite often. When a program is heavily loaded, it will reuse the
pthread_t faster than the stack profiler can notice, meaning that the CPU time
might end up being negative:

1. cached previous thread with same pthread_t id cpu time = 1000
2. new thread with same pthread_t id cpu time = 500
3. cpu time = 500 - 1000 = -500 🤦

To fix that, we now also uses the native thread id to identify thread. This
should be unique and the couple (pthread_t, native id) has a really small
chance of collision only using hash instead of native id (for old Python
versions).

We also make sure we never put a negative CPU time, if a collision ever
happens. The CPU time will just be 0 for the first (and maybe only) polling,
which is safer.

* feat(pprof): use thread_native_id to sort/group stack events

This make sure the thread native id is also used to identify threads in case
their (pthread_t id, name) are reused. This also embeds the thread native id as
a label which can be useful for debugging.