core: rewrite queue (#1534)
The queue implementation subclassed the stdlib Queue while using none of its
features besides the built-in lock. The notifying capabilities are not used.

It was also noticed that the queue put operation was strangely more expensive
than simply adding an item to a mutex protected list, possibly due to the
additional notifying capabilities which there is no need of.