feat: add startup logging (#1548)
* feat: add startup logging

A significant portion of our support cases can be solved using
information readily available to the library. This patch adds a module
`debug` which provides a function `collect` to query this information.

If the logging level is less than or equal to INFO then the information is
collected and logged at the end of `tracer.configure()`.