fix(sanic): handle invalid response type (#1647)
I caught this issue when running the sanic test suite with ddtrace enabled. A test meant to test the condition when an application route returns a str object rather than an HTTPResponse. The sanic integration only expected to handle the latter. This PR fixes this gap and includes a test for an async example.