fix(django): avoid mixing str and non-str args for uri helper (#1679)
With PY3 urlunparse calls urllib.parse._coerce_args which uses the
type of the scheme to check the type to expect from all url parts, raising
a TypeError otherwise. If the scheme is not a str, the function returns
the url parts bytes decoded along with a function to encode the result of
combining the url parts. We returns a byte string when all url parts are
byte strings.
https://github.com/python/cpython/blob/02d126aa09d96d03dcf9c5b51c858ce5ef386601/Lib/urllib/parse.py#L111-L125

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>