fix(django): add safe guards in get_request_uri (#1777)
This fix addresses a problem that arises if a request is handled that is missing a url scheme. An example of this case is seen in [saleor's wsgi code](https://github.com/mirumee/saleor/blob/a82166ce550eb4baefbc7ffb1c1514d710500f60/saleor/wsgi/__init__.py#L33-L45) where a warmup request is made when the wsgi application is started:

```
application(
    {
        "REQUEST_METHOD": "GET",
        "SERVER_NAME": SimpleLazyObject(get_allowed_host_lazy),
        "REMOTE_ADDR": "127.0.0.1",
        "SERVER_PORT": 80,
        "PATH_INFO": "/graphql/",
        "wsgi.input": b"",
        "wsgi.multiprocess": True,
    },
    lambda x, y: None,
)
```

Additionally, this fix addresses the use of a `SimpleLazyObject` for a request attribute.

I have added this to the [majorgreys/django-tutorial-uwsgi-warmup](https://github.com/DataDog/trace-examples/tree/majorgreys/django-tutorial-uwsgi-warmup) branch of the django-tutorial trace example. This can serve as a manual test to confirm the fix.