refactor(propagation): Make Propagator all static (#2101)
All the current implementations of the Propagator abstract class could
be made to have static methods. This avoids having to create objects
that are used just to invoke a method which could be called on the class
itself.

Only drawback is if, in the future, the inject and extract method would
depend on some internal state of the Propagator object.

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>